Move command line to its own file.
